[CODE REVIEW CORRECTS] refactored controllers
