+++ app/controllers/additional_codes/additional_code_types_controller.rb
@@ -1,5 +1,5 @@
-module Measures
-  class AdditionalCodeTypesController < ::Measures::BaseController
+module AdditionalCodes
+  class AdditionalCodeTypesController < ::BaseController
 
     def collection
       AdditionalCodeType.actual.q_search(params[:q])
