+++ app/controllers/certificates/certificate_types_controller.rb
@@ -1,0 +1,8 @@
+module Certificates
+  class CertificateTypesController < ::BaseController
+
+    def collection
+      CertificateType.actual
+    end
+  end
+end
