+++ app/controllers/certificates/certificates_controller.rb
@@ -1,0 +1,8 @@
+module Certificates
+  class CertificatesController < ::BaseController
+
+    def collection
+      Certificate.actual
+    end
+  end
+end
