+++ app/controllers/duty_expressions_controller.rb
@@ -1,0 +1,6 @@
+class DutyExpressionsController < ::BaseController
+
+  def collection
+    DutyExpression.actual
+  end
+end
