+++ app/controllers/footnotes/footnote_types_controller.rb
@@ -1,0 +1,8 @@
+module Footnotes
+  class FootnoteTypesController < ::BaseController
+
+    def collection
+      FootnoteType.actual
+    end
+  end
+end
