+++ app/controllers/footnotes/footnotes_controller.rb
@@ -1,10 +1,9 @@
-module Measures
-  class FootnotesController < ::Measures::BaseController
+module Footnotes
+  class FootnotesController < ::BaseController
 
     def collection
       Footnote.actual
               .q_search(params[:footnote_type_id], params[:description])
-              .limit(20)
     end
   end
 end
