--- app/controllers/measures/base_controller.rb
@@ -1,20 +1,0 @@
-module Measures
-  class BaseController < ApplicationController
-
-    respond_to :json
-
-    expose(:json_list) do
-      list = []
-
-      collection.map do |record|
-        list << record.json_mapping
-      end
-
-      list
-    end
-
-    def index
-      render json: json_list, status: :ok
-    end
-  end
-end
