--- app/controllers/measures/certificate_types_controller.rb
@@ -1,9 +1,0 @@
-module Measures
-  class CertificateTypesController < ::Measures::BaseController
-
-    def collection
-      CertificateType.actual
-                     .limit(100)
-    end
-  end
-end
