--- app/controllers/measures/certificates_controller.rb
@@ -1,9 +1,0 @@
-module Measures
-  class CertificatesController < ::Measures::BaseController
-
-    def collection
-      Certificate.actual
-                 .limit(100)
-    end
-  end
-end
