--- app/controllers/measures/duty_expressions_controller.rb
@@ -1,8 +1,0 @@
-module Measures
-  class DutyExpressionsController < ::Measures::BaseController
-
-    def collection
-      DutyExpression.actual
-    end
-  end
-end
