--- app/controllers/measures/footnote_types_controller.rb
@@ -1,8 +1,0 @@
-module Measures
-  class FootnoteTypesController < ::Measures::BaseController
-
-    def collection
-      FootnoteType.actual
-    end
-  end
-end
