--- app/controllers/measures/goods_nomenclatures_controller.rb
@@ -1,17 +1,0 @@
-module Measures
-  class GoodsNomenclaturesController < ApplicationController
-    respond_to :json
-
-    def index
-      @nomenclature = GoodsNomenclature.actual
-                                       .where(goods_nomenclature_item_id: params[:q])
-                                       .first.try(:sti_instance)
-
-      if @nomenclature.present?
-        render partial: "shared/tariff_breadcrumbs"
-      else
-        head :not_found
-      end
-    end
-  end
-end
