+++ app/controllers/measures/measurement_unit_qualifiers_controller.rb
@@ -1,5 +1,5 @@
 module Measures
-  class MeasurementUnitQualifiersController < ::Measures::BaseController
+  class MeasurementUnitQualifiersController < ::BaseController
 
     def collection
       MeasurementUnitQualifier.actual
