--- app/controllers/measures/regulations_controller.rb
@@ -1,11 +1,0 @@
-module Measures
-  class RegulationsController < ::Measures::BaseController
-
-    def collection
-      BaseRegulation.actual
-                    .not_replaced_and_partially_replaced
-                    .q_search(params[:q])
-                    .limit(100)
-    end
-  end
-end
