+++ app/controllers/regulations_controller.rb
@@ -1,0 +1,8 @@
+class RegulationsController < ::BaseController
+
+  def collection
+    BaseRegulation.actual
+                  .not_replaced_and_partially_replaced
+                  .q_search(params[:q])
+  end
+end
